stdlib.python.device: uint64 to avoid Python OverflowError

+++ +stdlib/+python/device.m
@@ -1,8 +1,9 @@
 function i = device(file)
 
 try
-  i = int64(py.os.stat(file).st_dev);
-  % int64 first is for Matlab <= R2022a
+  i = py.os.stat(file).st_dev;
+  % Matlab <= R2022a wants int64 before uint64, but this can make OverflowError on Windows within Python
+  % because on Windows st_dev is a 64-bit unsigned integer
 catch e
   pythonException(e)
   i = [];
